requirements.txt
sqlalchemy
pandas
psycopg2-binary==2.9.9
sqlalchemy
python-dotenv
